Added support for internal addresses
If no internal address is utilized, the address provided is used.
Internal addresses are used to specify the address, on which
a service is started. If specifie, the parameter provided with -a
will be used to communicate the address to the registry, while
-aint will be used to specify the address on which the socket
should listen to.

In case of the CoSimulationStandalone, an addition parameter for providing
the remote access service is now included.